no-mistakes(review): Harden Markdown rendering and stabilize bundle tests
