+++ package.json
@@ -48,7 +48,7 @@
     "eslint": "^10.3.0",
     "eslint-config-prettier": "^10.1.8",
     "globals": "^17.6.0",
-    "jsdom": "^29.1.1",
+    "jsdom": "26.1.0",
     "mermaid": "11.12.1",
     "prettier": "^3.8.3",
     "react": "18.3.1",
