+++ scripts/build.js
@@ -2,6 +2,8 @@ import { chmod, copyFile, cp, mkdir, readFile } from "node:fs/promises";
 
 import * as esbuild from "esbuild";
 
+import { buildChromeClient } from "./build-chrome-client.js";
+
 const packageJson = JSON.parse(await readFile(new URL("../package.json", import.meta.url), "utf8"));
 
 await mkdir("dist", { recursive: true });
@@ -25,14 +27,8 @@ await chmod("dist/cli.mjs", 0o755);
 
 // Chrome client: browser IIFE bundling marked + DOMPurify + chat-markdown so
 // /chrome-client.js stays a single self-contained asset for packaged and source runs.
-await esbuild.build({
-  entryPoints: ["src/chrome-client.js"],
+await buildChromeClient({
   outfile: "dist/chrome-client.js",
-  bundle: true,
-  minify: true,
-  format: "iife",
-  platform: "browser",
-  target: ["es2022"],
 });
 await copyFile("src/chrome.css", "dist/chrome.css");
 await mkdir("dist/design", { recursive: true });
